# fast path for dists in the PointMass constraint