Previously the PointMass form constraint called slow optimization routine for `<: Distribution` objects, even though it is easier and faster to return the `mode` of the distribution. This PR fixes that. Found by @bartvanerp