@bartvanerp 2023-08-28
Before:
```julia
BenchmarkTools.Trial: 6 samples with 1 evaluation.
 Range (min … max):  929.941 ms …   1.077 s  ┊ GC (min … max):  9.92% … 11.26%
 Time  (median):     957.682 ms              ┊ GC (median):    11.08%
 Time  (mean ± σ):   987.136 ms ± 64.681 ms  ┊ GC (mean ± σ):  11.12% ±  0.82%

  █  █      █ █                                        █     █  
  █▁▁█▁▁▁▁▁▁█▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁█ ▁
  930 ms          Histogram: frequency by time          1.08 s <

 Memory estimate: 528.79 MiB, allocs estimate: 4246814.
```

After
```julia
BenchmarkTools.Trial: 7 samples with 1 evaluation.
 Range (min … max):  668.098 ms … 845.756 ms  ┊ GC (min … max):  9.63% … 12.63%
 Time  (median):     725.226 ms               ┊ GC (median):    11.92%
 Time  (mean ± σ):   739.858 ms ±  55.859 ms  ┊ GC (mean ± σ):  11.65% ±  1.28%

  █             █   ██  █            █                        █  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁██▁▁█▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  668 ms           Histogram: frequency by time          846 ms <

 Memory estimate: 485.53 MiB, allocs estimate: 3522721.
```